+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Added
 - **Add ability to create multipage PDF files (N-up)** (based on podofoimpose)
-- [ ] Add ability to create custom multipage profiles
+- Add ability to create custom multipage profiles
 - Add live preview of output pages layout
 - Add a "Credits" tab to the about dialog
 - Add a "Contribute" tab to the about dialog